pyproject.toml
[project]
name = "gso-scaffolds"
version = "0.1.0"
description = "Example scaffolds for running agents on the GSO benchmark"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "gsobench",
    "datasets>=2.0.0",
]

[project.scripts]
gso-harbor = "harbor.cli:main"
gso-openhands = "openhands_gso.run_infer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harbor", "openhands_gso"]
